mybooks/requirements.txt
Django==3.1
psycopg2-binary==2.8.5
djangorestframework==3.11.1
python-dotenv==0.14.0
